requirements.txt
attrs==23.2.0
brother-ql==0.9.4
click==8.1.7
future==1.0.0
packbits==0.6
pillow==10.3.0
pyusb==1.2.1
